llm/requirements.txt
docling==2.37.0
fastapi==0.115.12
ollama==0.4.7
openai==1.87.0
pydantic==2.11.7
pymilvus==2.5.8
python-dotenv==1.1.0
python-multipart==0.0.20
Requests==2.32.4
tiktoken==0.9.0
transformers==4.51.3
uvicorn==0.34.1
pypdf==5.6.0